warnings fixed

+++ test/test_xbuffer_adaptor.cpp
@@ -31,7 +31,7 @@ namespace xt
         owner_adaptor adapt2(std::move(adapt));
         EXPECT_EQ(data, adapt2.data());
         EXPECT_EQ(size, adapt2.size());
-        EXPECT_EQ(0, adapt.size());
+        EXPECT_EQ(size_t(0), adapt.size());
     }
 
     TEST(xbuffer_adaptor, owner_copy_assign)
